# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Klyro API",
    "version": "0.1.0",
    "description": "Public scheduling API for the Klyro platform."
  },
  "servers": [
    {
      "url": "https://klyro.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/appointments/{id}/cancel": {
      "post": {
        "operationId": "post-appointments-{id}-cancel",
        "summary": "Cancel an appointment via a client-held cancel token.",
        "description": "Public endpoint — no session required. Authenticated via",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cancelAppointmentSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cancelResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (invalid UUID format)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "Appointment already in a terminal state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/booking/create": {
      "post": {
        "operationId": "post-booking-create",
        "summary": "Create a booking appointment",
        "description": "Books a slot for a client. Verifies availability, upserts the client by phone, and creates the appointment with a unique KLY-XXXX booking code.",
        "tags": [
          "Booking"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createBookingSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/bookingCreatedResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found or hidden by RLS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "Slot already taken (pre-check or DB race)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/booking/slots": {
      "get": {
        "operationId": "get-booking-slots",
        "summary": "List available slots for a staff member on a date",
        "description": "Returns time slots that are open for booking given a staff member, service, branch, and local date.",
        "tags": [
          "Booking"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "staffId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available booking slots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/slotsListResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}/messages": {
      "get": {
        "operationId": "get-appointments-{id}-messages",
        "summary": "Get delivery status for all messages linked to an appointment.",
        "description": "Returns message rows for the given appointment. Session required;",
        "tags": [
          "Messaging"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Message statuses retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageStatusResponseSchema"
                }
              }
            }
          },
          "401": {
            "description": "No authenticated session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found or not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponseSchema"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Appointments"
    },
    {
      "name": "Booking"
    },
    {
      "name": "Messaging"
    }
  ],
  "components": {
    "schemas": {
      "bookingCreatedResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "bookingCode": {
                "type": "string",
                "pattern": "^KLY-[A-Z0-9]{4}$",
                "example": "KLY-7F4A"
              },
              "startsAt": {
                "type": "string",
                "format": "date-time",
                "example": "2026-05-26T15:00:00.000Z"
              },
              "endsAt": {
                "type": "string",
                "format": "date-time",
                "example": "2026-05-26T15:30:00.000Z"
              }
            },
            "required": [
              "bookingCode",
              "startsAt",
              "endsAt"
            ]
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "$ref": "#/components/schemas/linkSchema"
              }
            },
            "required": [
              "self"
            ]
          }
        },
        "required": [
          "data",
          "_links"
        ]
      },
      "cancelAppointmentSchema": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "token"
        ]
      },
      "cancelResponseSchema": {},
      "createBookingSchema": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "slotStart": {
            "type": "string",
            "format": "date-time"
          },
          "clientName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "clientPhone": {
            "type": "string",
            "pattern": "^\\+\\d{7,15}$"
          },
          "clientEmail": {
            "type": "string",
            "format": "email"
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "staffId",
          "serviceId",
          "branchId",
          "slotStart",
          "clientName",
          "clientPhone"
        ]
      },
      "errorResponseSchema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "VALIDATION_ERROR",
                  "SLOT_TAKEN",
                  "BOOKING_CONFLICT",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "field": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "linkSchema": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PATCH",
              "PUT",
              "DELETE"
            ]
          }
        },
        "required": [
          "href",
          "method"
        ]
      },
      "messageStatusResponseSchema": {},
      "slotsListResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "startsAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2026-05-26T15:00:00.000Z"
                },
                "endsAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2026-05-26T15:30:00.000Z"
                }
              },
              "required": [
                "startsAt",
                "endsAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991,
                "example": 8
              }
            },
            "required": [
              "total"
            ]
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "$ref": "#/components/schemas/linkSchema"
              }
            },
            "required": [
              "self"
            ]
          }
        },
        "required": [
          "data",
          "meta",
          "_links"
        ]
      },
      "slotsQuerySchema": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "string",
            "format": "uuid"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          }
        },
        "required": [
          "staffId",
          "serviceId",
          "branchId",
          "date"
        ]
      }
    },
    "responses": {}
  }
}